Revert "18 iss 04 create booking model"
