+++ backend/app/models/__init__.py
@@ -15,15 +15,10 @@ initialization.
 - `Notification`: Represents a system alert or message.
 """
 
-from .booking import (
-    Booking,
-    BookingStatus,
-    RecurrenceFrequency,
-    TimeSlot,
-    TimeslotStatus,
-)
+from .booking import Booking
 from .notification import Notification
 from .room import Room
+from .time_slot import TimeSlot
 from .user import User, UserRole
 
 __all__ = [
@@ -31,9 +26,6 @@ __all__ = [
     "UserRole",
     "Room",
     "TimeSlot",
-    "TimeslotStatus",
-    "RecurrenceFrequency",
-    "BookingStatus",
     "Booking",
     "Notification",
 ]
